requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
color-pprint==0.0.1
ezcord
colorama==0.4.6
discord-ui==5.1.6
discord.py @ git+https://github.com/Rapptz/discord.py@2fdbe59376d736483cd1226e674e609433877af4
frozenlist==1.3.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
psutil==5.9.5
pycountry==22.3.5
pycryptodome==3.18.0
pyfiglet==0.8.post1
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4.1
typing_extensions==4.5.0
urllib3==2.0.2
chat_exporter
yarl==1.9.2
scrapetube
openai
pywin32==306
